add coords including trench ends
Adds x and y coordinates to a layer without creating a new virtual layer. Create a new shapefile or modify the existing one. This is for archaeological trenching, so it only does the start and end of lines and has an option to add the midpoint of the 2 narrowest sides of a rectangle polygon. Polygons that are not rectangles will not get accurate coordinates for the start and end points