Add files via upload
Processing toolbox script.  Creates sequential user-defined buffers, then uses the DIFFERENCE tool to turn them into donuts. Each buffer is saved as a new layer in the project home /donut_buffers for ease of (and more options for) query and display. The solid buffers are saved in a sub-folder /donut_buffers/gluten_free.